--- crates/lib/docs_rs_crate_zip/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "docs_rs_crate_zip"
-description = "read specific files from the crates.io source zip archive"
-version = "0.1.0"
-edition = "2024"
-
-[features]
-testing = [
-    "dep:zip",
-    "dep:mockito",
-    "dep:serde_json",
-]
-
-[dependencies]
-anyhow = { workspace = true }
-async-compression = { workspace = true }
-docs_rs_utils = { path = "../docs_rs_utils" }
-futures-util = { workspace = true }
-reqwest = { workspace = true }
-serde = { workspace = true }
-tokio = { workspace = true }
-tokio-util = { workspace = true }
-zip = { workspace = true, optional = true }
-mockito = { workspace = true, optional = true }
-serde_json = { workspace = true, optional = true }
-
-[dev-dependencies]
-zip = { workspace = true }
-mockito = { workspace = true }
-tokio = { workspace = true, features = ["io-std"] }
-serde_json = { workspace = true }
