+++ crates/lib/docs_rs_registry_api/Cargo.toml
@@ -5,25 +5,34 @@ license = "MIT"
 repository = "https://github.com/rust-lang/docs.rs"
 edition = "2024"
 
+[features]
+testing = [
+    "dep:mockito",
+    "dep:zip",
+]
+
 [dependencies]
 anyhow = { workspace = true }
+async-compression = { workspace = true }
 bon = { workspace = true }
 chrono = { workspace = true }
 docs_rs_config = { path = "../docs_rs_config" }
 docs_rs_env_vars = { path = "../docs_rs_env_vars" }
 docs_rs_types = { path = "../docs_rs_types" }
 docs_rs_utils = { path = "../docs_rs_utils" }
-tokio = { workspace = true }
 futures-util = { workspace = true }
-async-compression = { workspace = true }
-tokio-util = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 sqlx = { workspace = true }
 thiserror = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
+mockito = { workspace = true, optional = true }
+zip = { workspace = true, optional = true }
+
 
 [dev-dependencies]
 docs_rs_types = { path = "../docs_rs_types", features = ["testing"] }
@@ -31,6 +40,7 @@ mime = { workspace = true }
 mockito = { workspace = true }
 test-case = { workspace = true }
 tokio = { workspace = true }
+zip = { workspace = true }
 
 [lints]
 workspace = true
