+++ crates/lib/docs_rs_registry_api/src/lib.rs
@@ -4,7 +4,14 @@ mod error;
 mod models;
 mod source_archive;
 
+#[cfg(any(test, feature = "testing"))]
+pub mod testing;
+
 pub use api::RegistryApi;
 pub use config::Config;
 pub use error::Error;
 pub use models::{CrateData, CrateOwner, OwnerKind, ReleaseData, Search};
+pub use source_archive::{
+    SourceArchive,
+    manifest::{FileEntry, Manifest},
+};
