+++ crates/lib/docs_rs_registry_api/src/testing/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod static_test_env;
